+++ addon/services/esri-loader.js
@@ -15,6 +15,12 @@ import esriLoader from 'esri-loader';
 
 export default Ember.Service.extend({
 
+  init () {
+    this._super(...arguments);
+    // have esriLoader use Ember's RSVP promise
+    esriLoader.utils.Promise = Ember.RSVP.Promise;
+  },
+
   // emulate computed property isLoaded to indicate that the JSAPI been loaded
   unknownProperty (key) {
     if (key === 'isLoaded') {
@@ -22,94 +28,31 @@ export default Ember.Service.extend({
     }
   },
 
+  loadScript (options) {
+    return esriLoader.loadScript(options)
+    .then(script => {
+      // have to wrap this async side effect in Ember.run() so tests don't fail
+      // may be able to remove this once we can have esriLoader use RSVP.Promise
+      Ember.run(() => {
+        // update the isLoaded computed property
+        this.notifyPropertyChange('isLoaded');
+        return script;
+      });
+    });
+  },
+
   // inject a script tag pointing to the JSAPI in the page
   // and return a promise once it's loaded
   load (options = {}) {
-    // if already loading or loaded, return the existing promise
-    if (this._loadPromise) {
-      return this._loadPromise;
-    }
-    // otherwise create a promise that will resolve when the JSAPI is loaded
-    this._loadPromise = new Ember.RSVP.Promise((resolve, reject) => {
-      esriLoader.bootstrap(err => {
-        if (err) {
-          const message = err.message || err;
-          // has esriLoader.bootstrap already been called?
-          if (message === 'The ArcGIS API for JavaScript is already loaded.') {
-            // this can happen when there is more than one instance of this service
-            // running on the page at a time, for example, in acceptance tests
-
-            // TODO: this is _much_ better handled upstream
-            // so we want to get rid of all this once this issue is resolved:
-            // https://github.com/Esri/esri-loader/issues/28
-
-            // first check if it's the same script
-            // NOTE: will haev to update this every time it's updated here:
-            // https://github.com/Esri/esri-loader/blob/master/src/esri-loader.ts#L29
-            const defaultUrl = 'https://js.arcgis.com/4.4/';
-            const url = options.url || defaultUrl;
-            const script = document.querySelector('script[data-esri-loader]');
-            if (script.src !== url) {
-              // user tried to load two different JSAPI scripts
-              reject(err);
-            } else {
-              // check if the script has loaded yet
-              if (script.dataset.esriLoader === 'loaded' || esriLoader.isLoaded()) {
-                resolve();
-              } else {
-                // wait for the script to load and then resolve
-                script.addEventListener('load', () => {
-                  // TODO: remove this event listener
-                  resolve();
-                }, false);
-              }
-            }
-          } else {
-            // not an error we can handle
-            reject(err);
-          }
-        } else {
-          // no err
-          resolve();
-        }
-      }, options);
-    })
-    .then(() => {
-      // update the isLoaded computed property
-      this.notifyPropertyChange('isLoaded');
-      // let the caller know that the API has been successfully loaded
-      // TODO: would there be something more useful to return here?
-      // bootstrap returns dojoRequire,
-      // but we want consumers to use loadModules instead
-      return { success: true };
+    Ember.deprecate('esriLoader.load() will be removed at the next breaking version. Use esriLoader.loadScript() instead.', false, {
+      id: 'ember-esri-loader.load',
+      until: '10.0.0'
     });
-    return this._loadPromise;
+    return this.loadScript(options);
   },
 
   // require the modules and return a pomise that reolves them as an array
-  loadModules (moduleNames) {
-    // TODO: validate that moduleNames is an array w/ at least one string?
-    // or just continue to let dojo throw "Cannot read property 'has' of undefined"?
-    // TODO: we can probably delegate some or all of this logic to esriLoader.dojoRequire
-    if (this.get('isLoaded')) {
-      return this._loadModules(moduleNames);
-    } else {
-      if (this._loadPromise) {
-        // load modules once finished loadng the JSAPI
-        return this._loadPromise.then(Ember.run.bind(this, this._loadModules, moduleNames));
-      } else {
-        // not loaded or loading
-        return Ember.RSVP.reject(new Error('The ArcGIS API for JavaScript has not been loaded. You must first call esriLoader.load()'));
-      }
-    }
-  },
-
-  // wrap esriLoader's dojoRequire in a promise
-  _loadModules (moduleNames) {
-    return new Ember.RSVP.Promise(resolve => {
-      esriLoader.dojoRequire(moduleNames, (...modules) => {
-        resolve(modules);
-      });
-    });
+  loadModules (moduleNames, options) {
+    return esriLoader.loadModules(moduleNames, options);
   }
 });